# frontend/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenApi specification","description":"OpenApi documentation for spring security","termsOfService":"terms of service","contact":{"name":"elaa","url":"https://elaBrahmi.com","email":"elaa.brahmii@gmail.com"},"license":{"name":"license name","url":"https://some-url.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8088/api/v1","description":"local env"},{"url":"http://other-url","description":"prod env"}],"security":[{"bearerAuth":[]}],"paths":{"/user":{"get":{"tags":["user-controller"],"operationId":"getUserById","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/tasks/{idTask}":{"put":{"tags":["task"],"operationId":"updateTask","parameters":[{"name":"idTask","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}},"delete":{"tags":["task"],"operationId":"deleteTask","parameters":[{"name":"idTask","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/tasks/status/{idTask}":{"put":{"tags":["task"],"operationId":"updateTaskStatus","parameters":[{"name":"idTask","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskStatusRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/notification/{id}/read":{"put":{"tags":["notification"],"operationId":"markRead","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/tasks":{"get":{"tags":["task"],"operationId":"findAllTasksByUserId","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":6}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseTaskResponse"}}}}}},"post":{"tags":["task"],"operationId":"saveTask","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/auth/register":{"post":{"tags":["Authentication"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"202":{"description":"Accepted","content":{"application/json":{"schema":{"type":"object"}}}}}}},"/auth/authenticate":{"post":{"tags":["Authentication"],"operationId":"authenticate","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/tasks/status/{status}":{"get":{"tags":["task"],"operationId":"findTasksByStatus","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":6}},{"name":"status","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseTaskResponse"}}}}}}},"/tasks/search":{"get":{"tags":["task"],"operationId":"findTasksByKeyword","parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":6}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseTaskResponse"}}}}}}},"/tasks/priority/{priority}":{"get":{"tags":["task"],"operationId":"findTasksByPriority","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":6}},{"name":"priority","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseTaskResponse"}}}}}}},"/tasks/id/{task-id}":{"get":{"tags":["task"],"operationId":"findTaskById","parameters":[{"name":"task-id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}}}}},"/tasks/finished":{"get":{"tags":["task"],"operationId":"nbTasksFinished","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/tasks/category/{category}":{"get":{"tags":["task"],"operationId":"findTasksBycategory","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":6}},{"name":"category","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseTaskResponse"}}}}}}},"/tasks/all":{"get":{"tags":["task"],"operationId":"nbTasks","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int32"}}}}}}},"/tasks/DueDate/{DueDate}":{"get":{"tags":["task"],"operationId":"findTasksByDueDate","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":6}},{"name":"DueDate","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageResponseTaskResponse"}}}}}}},"/notification":{"get":{"tags":["notification"],"operationId":"getAllNotificationsByUser","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NotificationResponse"}}}}}}}},"/auth/activate-account":{"get":{"tags":["Authentication"],"operationId":"confirm","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UserRequest":{"required":["email","firstName","lastName"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"}}},"TaskRequest":{"required":["category","description","dueDate","status","title"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"dueDate":{"type":"string"},"category":{"type":"string"},"importance":{"type":"string","enum":["HIGH","MEDIUM","LOW"]}}},"TaskStatusRequest":{"type":"object","properties":{"status":{"type":"string"}}},"RegistrationRequest":{"required":["email","firstName","lastName","password"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthenticationRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"token":{"type":"string"}}},"UserResponse":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"}}},"PageResponseTaskResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/TaskResponse"}},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"},"first":{"type":"boolean"},"last":{"type":"boolean"}}},"TaskResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"title":{"type":"string"},"description":{"type":"string"},"category":{"type":"string"},"status":{"type":"string","enum":["FINISHED","IN_PROGRESS","PENDING"]},"assignee":{"type":"integer","format":"int32"},"importance":{"type":"string","enum":["HIGH","MEDIUM","LOW"]},"duedate":{"type":"string","format":"date"}}},"NotificationResponse":{"required":["createdAt","message"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"message":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["READ","UNREAD"]}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"JWT auth description","in":"header","scheme":"bearer","bearerFormat":"JWT"}}}}